GQL-109: Rename graphdb data source to graphDbRelatedCollections
