+++ src/graphql/handler.js
@@ -37,7 +37,7 @@ import dataQualitySummarySource from '../datasources/dataQualitySummary'
 import granuleSource from '../datasources/granule'
 import graphDbAssociatedCitations from '../datasources/graphDbAssociatedCitations'
 import graphDbDuplicateCollectionsSource from '../datasources/graphDbDuplicateCollections'
-import graphDbSource from '../datasources/graphDb'
+import graphDbRelatedCollectionsSource from '../datasources/graphDbRelatedCollections'
 import gridSource from '../datasources/grid'
 import maxItemsPerOrderSource from '../datasources/maxItemsPerOrder'
 import permissionSource from '../datasources/permission'
@@ -249,7 +249,7 @@ export default startServerAndCreateLambdaHandler(
           granuleSource,
           graphDbAssociatedCitations,
           graphDbDuplicateCollectionsSource,
-          graphDbSource,
+          graphDbRelatedCollectionsSource,
           gridSource,
           groupSourceCreate,
           groupSourceDelete,
